+++ client/src/main/java/com/chanshiyu/client/console/ListGroupMembersConsoleCommand.java
@@ -16,7 +16,7 @@ public class ListGroupMembersConsoleCommand implements ConsoleCommand {
     public void exec(Scanner scanner, Channel channel) {
         System.out.print("输入 groupId，获取群成员列表：");
         ListGroupMembersRequestPacket listGroupMembersRequestPacket = new ListGroupMembersRequestPacket();
-        long groupId = scanner.nextLong();
+        int groupId = scanner.nextInt();
         listGroupMembersRequestPacket.setGroupId(groupId);
         channel.writeAndFlush(listGroupMembersRequestPacket);
     }
