+++ client/src/main/java/com/chanshiyu/client/console/SendToUserConsoleCommand.java
@@ -15,7 +15,7 @@ public class SendToUserConsoleCommand implements ConsoleCommand {
     @Override
     public void exec(Scanner scanner, Channel channel) {
         System.out.print("发送消息给某个用户：");
-        long toUserId = scanner.nextLong();
+        int toUserId = scanner.nextInt();
         String message = scanner.next();
         channel.writeAndFlush(new MessageRequestPacket(toUserId, 1, message));
     }
