+++ client/src/main/java/com/chanshiyu/client/handler/response/GroupMessageResponseHandler.java
@@ -14,7 +14,7 @@ public class GroupMessageResponseHandler extends SimpleChannelInboundHandler<Gro
 
     @Override
     protected void channelRead0(ChannelHandlerContext ctx, GroupMessageResponsePacket responsePacket) {
-        long fromGroupId = responsePacket.getFromGroupId();
+        int fromGroupId = responsePacket.getFromGroupId();
         Session fromUser = responsePacket.getFromUser();
         System.out.println("收到群[" + fromGroupId + "]中[" + fromUser + "]发来的消息：" + responsePacket.getMessage());
     }
